Teensy integration
Partial design. Several other traces to go.
I'm using the strategy of leaving the old uC connected untill all the
traces have been connected to the Teensy. Also I'm connecting one trace
at the time to reduce tangling and crossing (with the constraint of
connecting IIC and analog pins to the hardware-dedicated ones. Two
issues remaining. 1)cope with the fact that the teensy internal
reference reference is 1.2V 2) thinking about a possibile better
filtering of the 3.3V analog supply.