Reached a dead end
Attempt to modify the circuit to obtain a faster filtering. The transfer
function from the fast amplifier to the monitor node has some poles that
would reduce margin of phase... Given the particoular criticality of the
circuit (The LM7171 has a very large bandwidth) the configuration using
LC will likely oscillate.  I have added the file
simulation_hv_filter.sch that is in QUCS format and allows to play with
transfer funtions. The decision is to hold on the LCR design.: It will
likely not work. A better way to increase the bandwidth is to trade
between noise performance and resistors' value in the current RC
configuration.